requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
python-multipart
stripe
passlib[bcrypt]
python-jose[cryptography]
